u128 -> u8

+++ assets/2025-04-06_No_Adjacent_Characters/rust/src/a1_simpletons_array.rs
@@ -1,4 +1,4 @@
-pub fn run(input: &mut [u128]) {
+pub fn run(input: &mut [u8]) {
     let len = input.len();
 
     if len < 2 {
